Proper strategy for merging streams
Don't try to let Spark do it, just do it in the application using cached
results of changed pages. This implementation avoids shuffles and
drastically cuts back on bytes written. Only potential issue is that
large sets of longs could be transported into the application if a very
large number of pages has been updated.